feat(frames) finish modularizing frame.rs into individual frame modules
